feat: Complete model and results with some nice text. Add code for latin text
The model generates some nice results and also the training effects can be seen,
how the model gets better and better.
The model had to be restarted, as after training, the kernel got stuck.

As a preparation, I also added code to read the latin text. But no tests and
model building yet.